requirements.txt
aenum==3.1.15
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.6.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
backports.zoneinfo==0.2.1
blinker==1.7.0
certifi==2023.11.17
channels==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Deprecated==1.2.14
Django==3.2
django-chartjs==2.3.0
django-crispy-forms==1.13.0
django-filter==23.3
django-fontawesome==1.0
django-fontawesome-5==1.0.18
django-plotly-dash==2.2.2
dpd_components==0.1.0
Flask==3.0.0
fontawesomefree==6.4.0
frozenlist==1.4.0
future==0.18.3
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==1.18.0
line_notify==0.1.4
MarkupSafe==2.1.3
multidict==6.0.4
mysqlclient==2.2.0
ngrok==0.12.1
numpy==1.24.4
packaging==23.2
pandas==2.0.3
Pillow==9.0.0
plotly==5.18.0
pydantic==2.5.2
pydantic_core==2.14.5
PyMySQL==1.1.0
pypng==0.20220715.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.2
tenacity==8.2.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
